+++ ledger/sandbox-perf/src/perf/lib/scala/com/digitalasset/platform/sandbox/perf/LedgerFactories.scala
@@ -6,8 +6,6 @@ package com.digitalasset.platform.sandbox.perf
 import java.io.File
 
 import akka.stream.Materializer
-import ch.qos.logback.classic.Level
-import com.daml.ledger.participant.state.v1.TimeModel
 import com.digitalasset.daml.lf.archive.UniversalArchiveReader
 import com.digitalasset.daml.lf.data.Ref
 import com.digitalasset.grpc.adapter.ExecutionSequencerFactory
@@ -15,7 +13,6 @@ import com.digitalasset.ledger.api.domain.LedgerId
 import com.digitalasset.ledger.api.testing.utils.Resource
 import com.digitalasset.platform.common.LedgerIdMode
 import com.digitalasset.platform.sandbox.config.SandboxConfig
-import com.digitalasset.platform.services.time.TimeProviderType
 import com.digitalasset.testing.postgresql.{PostgresFixture, PostgresResource}
 
 object LedgerFactories {
@@ -24,23 +21,12 @@ object LedgerFactories {
     UniversalArchiveReader().readFile(file).map(_.all.head._1).get
 
   private def sandboxConfig(jdbcUrl: Option[String], darFiles: List[File]) =
-    SandboxConfig(
-      address = None,
+    SandboxConfig.default.copy(
       port = 0,
-      portFile = None,
       damlPackages = darFiles,
-      timeProviderType = TimeProviderType.Static,
-      timeModel = TimeModel.reasonableDefault,
-      commandConfig = SandboxConfig.defaultCommandConfig,
-      scenario = None,
-      tlsConfig = None,
       ledgerIdMode =
         LedgerIdMode.Static(LedgerId(Ref.LedgerString.assertFromString("ledger-server"))),
-      maxInboundMessageSize = SandboxConfig.DefaultMaxInboundMessageSize,
       jdbcUrl = jdbcUrl,
-      eagerPackageLoading = false,
-      logLevel = Level.INFO,
-      authService = None,
     )
 
   val mem = "InMemory"
