+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/apiserver/InMemoryPackageStore.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2020 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.platform.sandbox.stores
+package com.digitalasset.platform.apiserver
 
 import java.io.File
 import java.time.Instant
