+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/apiserver/JdbcIndex.scala
@@ -13,8 +13,6 @@ import com.daml.ledger.participant.state.v1.{ParticipantId, ReadService}
 import com.digitalasset.dec.DirectExecutionContext
 import com.digitalasset.ledger.api.domain.{ParticipantId => _, _}
 import com.digitalasset.logging.LoggingContext
-import com.digitalasset.platform.sandbox.stores.LedgerBackedIndexService
-import com.digitalasset.platform.sandbox.stores.ledger.{MeteredReadOnlyLedger, SandboxContractStore}
 import com.digitalasset.resources.Resource
 
 object JdbcIndex {
