+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/sandbox/stores/ledger/ScenarioLoader.scala
@@ -14,8 +14,9 @@ import com.digitalasset.daml.lf.transaction.{GenTransaction, VersionTimeline}
 import com.digitalasset.daml.lf.types.Ledger.ScenarioTransactionId
 import com.digitalasset.daml.lf.types.{Ledger => L}
 import com.digitalasset.daml.lf.value.Value.AbsoluteContractId
+import com.digitalasset.platform.apiserver.InMemoryPackageStore
 import com.digitalasset.platform.index.store.entries.LedgerEntry
-import com.digitalasset.platform.sandbox.stores.{InMemoryActiveLedgerState, InMemoryPackageStore}
+import com.digitalasset.platform.sandbox.stores.InMemoryActiveLedgerState
 import org.slf4j.LoggerFactory
 
 import scala.annotation.tailrec
